requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
pandas==2.1.4
plotly==5.18.0
gunicorn==21.2.0
openpyxl==3.1.2
matplotlib==3.8.2
matplotlib-venn==0.11.9
python-dateutil==2.8.2